Join all annotation extensions in one file
